Nchalo Water Project – Household Survey 2021

This is a community-level household survey conducted as part of the Nchalo Water Project. Its purpose is to assess livelihoods, water usage, community governance, infrastructure access, and opportunities for value addition in rural water supply and WASH-related development intervention.

Usage

nchalohhwatersurvey

Format

A tibble with 74 rows and 39 variables

submitted_on

Date and time the survey was submitted

community_name

Name of the community surveyed

latitude

Latitude coordinate of the survey location

longitude

Longitude coordinate of the survey location

community_population

Estimated total number of people in the community

main_livelihood_source

Primary source of livelihood for the household

main_livelihood_source_comments

Additional comments on the households main livelihood

income_in_2020_mwk

Estimated household income in 2020 from the main livelihood source (in MWK)

income_comments

Additional comments about the reported income

income_spending_priority

Main priorities for spending the household income

main_drinking_water_source

Primary source of drinking water for the community

main_drinking_water_source_other

Specification if Other water source was selected

uses_borehole_for_other_purposes

Whether the borehole is used for purposes other than drinking

other_borehole_use_comments

Comments on other uses of borehole water

reasons_not_pursuing_multiple_uses

Reasons for not using borehole water for multiple purposes

reasons_not_pursuing_multiple_uses_other

Specification if an Other reason was selected

value_addition_opportunity

Whether value addition opportunities exist for a new water project

value_addition_opportunity_details

Details about the identified value addition opportunities

value_addition_opportunity_other

Specification if an Other opportunity was mentioned

land_allocation_responsible_person

Person or group responsible for land allocation for new projects

leadership_allocation_decider

Entity that decides leadership roles in community projects

leadership_allocation_other

Specification if an Other decision maker was indicated

gender_guidelines_for_leadership

Existence of gender guidelines in community leadership

gender_guidelines_comments

Comments on the gender guidelines

pay_for_water

Whether the community pays for water usage

water_payment_amount

Amount paid for water usage (in MWK)

water_payment_frequency

Frequency of water payment

water_payment_frequency_other

Specification if an Other frequency was selected

reason_for_not_paying

Main reason for not paying for water

reason_for_not_paying_other

Specification if an Other reason was given

nearest_school_type

Type of school closest to the community

govt_school_accessibility

Whether government schools are accessible to all

nearest_health_facility_type

Type of health facility closest to the community

health_facility_accessibility

Whether the health facility is affordable and accessible to all

health_facility_accessibility_other

Specification if an Other accessibility concern was mentioned

community_solved_social_problem

Whether the community has independently addressed social problems

community_solved_social_problem_comments

Comments on how social problems were addressed by the community

flow_rate_test_possible_today

Whether it was possible to conduct a flow rate test during the visit

has_temp_ec_tds_meter

Whether the team had a temperature and EC or TDS meter available